Backend/requirements.txt
fastapi 
sqlalchemy 
psycopg2 
pydantic 
alembic
python-dotenv
asyncpg 
uvicorn
supabase

